Refactor for render posts
